Added test for ReviewLoaderStatic

+++ src/main/java/com/chdryra/android/reviewer/PlugIns/PersistencePlugin/PersistenceDatabase/PersistenceReviewerDb/Implementation/ReviewDataHolderImpl.java
@@ -86,7 +86,7 @@ public class ReviewDataHolderImpl implements ReviewDataHolder{
     }
 
     @Override
-    public float getRatingWeight() {
+    public int getRatingWeight() {
         return mRatingWeight;
     }
